+++ src/Math/Color.cpp
@@ -9,13 +9,13 @@
 
 using namespace Raycer;
 
-const Color Color::RED = Color(1.0f, 0.0f, 0.0f);
-const Color Color::GREEN = Color(0.0f, 1.0f, 0.0f);
-const Color Color::BLUE = Color(0.0f, 0.0f, 1.0f);
-const Color Color::WHITE = Color(1.0f, 1.0f, 1.0f);
-const Color Color::BLACK = Color(0.0f, 0.0f, 0.0f);
+const Color Color::RED = Color(1.0, 0.0, 0.0);
+const Color Color::GREEN = Color(0.0, 1.0, 0.0);
+const Color Color::BLUE = Color(0.0, 0.0, 1.0);
+const Color Color::WHITE = Color(1.0, 1.0, 1.0);
+const Color Color::BLACK = Color(0.0, 0.0, 0.0);
 
-Color::Color(float r_, float g_, float b_, float a_) : r(r_), g(g_), b(b_), a(a_)
+Color::Color(double r_, double g_, double b_, double a_) : r(r_), g(g_), b(b_), a(a_)
 {
 }
 
@@ -23,7 +23,7 @@ Color::Color(int r_, int g_, int b_, int a_)
 {
 	assert(r_ >= 0 && r_ <= 255 && g_ >= 0 && g_ <= 255 && b_ >= 0 && b_ <= 255 && a_ >= 0 && a_ <= 255);
 
-	float inv255 = 1.0f / 255.0f;
+	double inv255 = 1.0 / 255.0;
 
 	r = r_ * inv255;
 	g = g_ * inv255;
@@ -38,7 +38,7 @@ Color::Color(uint32_t abgr)
 	int b_ = (abgr >> 16) & 0xff;
 	int a_ = (abgr >> 24);
 
-	float inv255 = 1.0f / 255.0f;
+	double inv255 = 1.0 / 255.0;
 
 	r = r_ * inv255;
 	g = g_ * inv255;
@@ -48,12 +48,12 @@ Color::Color(uint32_t abgr)
 
 uint32_t Color::getAbgrValue() const
 {
-	assert(r >= 0.0f && r <= 1.0f && g >= 0.0f && g <= 1.0f && b >= 0.0f && b <= 1.0f && a >= 0.0f && a <= 1.0f);
+	assert(r >= 0.0 && r <= 1.0 && g >= 0.0 && g <= 1.0 && b >= 0.0 && b <= 1.0 && a >= 0.0 && a <= 1.0);
 
-	int ri = (int)(r * 255.0f + 0.5f);
-	int gi = (int)(g * 255.0f + 0.5f);
-	int bi = (int)(b * 255.0f + 0.5f);
-	int ai = (int)(a * 255.0f + 0.5f);
+	int ri = (int)(r * 255.0 + 0.5);
+	int gi = (int)(g * 255.0 + 0.5);
+	int bi = (int)(b * 255.0 + 0.5);
+	int ai = (int)(a * 255.0 + 0.5);
 
 	return (ai << 24 | bi << 16 | gi << 8 | ri);
 }
@@ -73,17 +73,17 @@ Color Raycer::operator*(const Color& c1, const Color& c2)
 	return Color(c1.r * c2.r, c1.g * c2.g, c1.b * c2.b, c1.a * c2.a);
 }
 
-Color Raycer::operator*(const Color& c, float s)
+Color Raycer::operator*(const Color& c, double s)
 {
 	return Color(c.r * s, c.g * s, c.b * s, c.a * s);
 }
 
-Color Raycer::operator*(float s, const Color& c)
+Color Raycer::operator*(double s, const Color& c)
 {
 	return Color(c.r * s, c.g * s, c.b * s, c.a * s);
 }
 
-Color Raycer::operator/(const Color& c, float s)
+Color Raycer::operator/(const Color& c, double s)
 {
 	return Color(c.r / s, c.g / s, c.b / s, c.a / s);
 }
@@ -112,14 +112,14 @@ Color& Color::operator-=(const Color& c)
 	return *this;
 }
 
-Color& Color::operator*=(float s)
+Color& Color::operator*=(double s)
 {
 	*this = *this * s;
 
 	return *this;
 }
 
-Color& Color::operator/=(float s)
+Color& Color::operator/=(double s)
 {
 	*this = *this / s;
 
@@ -128,7 +128,7 @@ Color& Color::operator/=(float s)
 
 bool Color::isTransparent() const
 {
-	return (a < 1.0f);
+	return (a < 1.0);
 }
 
 void Color::clamp()
@@ -140,15 +140,15 @@ Color Color::clamped() const
 {
 	Color c;
 
-	c.r = std::max(0.0f, std::min(r, 1.0f));
-	c.g = std::max(0.0f, std::min(g, 1.0f));
-	c.b = std::max(0.0f, std::min(b, 1.0f));
-	c.a = std::max(0.0f, std::min(a, 1.0f));
+	c.r = std::max(0.0, std::min(r, 1.0));
+	c.g = std::max(0.0, std::min(g, 1.0));
+	c.b = std::max(0.0, std::min(b, 1.0));
+	c.a = std::max(0.0, std::min(a, 1.0));
 
 	return c;
 }
 
-Color Color::lerp(const Color& start, const Color& end, float alpha)
+Color Color::lerp(const Color& start, const Color& end, double alpha)
 {
 	Color c;
 
@@ -162,15 +162,15 @@ Color Color::lerp(const Color& start, const Color& end, float alpha)
 
 Color Color::alphaBlend(const Color& first, const Color& second)
 {
-	float alpha = second.a;
-	float invAlpha = 1.0f - alpha;
+	double alpha = second.a;
+	double invAlpha = 1.0 - alpha;
 
 	Color c;
 
 	c.r = (alpha * second.r + invAlpha * first.r);
 	c.g = (alpha * second.g + invAlpha * first.g);
 	c.b = (alpha * second.b + invAlpha * first.b);
-	c.a = 1.0f;
+	c.a = 1.0;
 
 	return c;
 }
