+++ src/Math/Color.h
@@ -11,31 +11,31 @@ namespace Raycer
 	{
 	public:
 
-		Color(float r = 0.0f, float g = 0.0f, float b = 0.0f, float a = 1.0f);
+		Color(double r = 0.0, double g = 0.0, double b = 0.0, double a = 1.0);
 		explicit Color(int r, int g, int b, int a = 255);
 		Color(uint32_t abgr);
 
 		friend Color operator+(const Color& c1, const Color& c2);
 		friend Color operator-(const Color& c1, const Color& c2);
 		friend Color operator*(const Color& c1, const Color& c2);
-		friend Color operator*(const Color& c, float s);
-		friend Color operator*(float s, const Color& c);
-		friend Color operator/(const Color& c, float s);
+		friend Color operator*(const Color& c, double s);
+		friend Color operator*(double s, const Color& c);
+		friend Color operator/(const Color& c, double s);
 
 		friend bool operator==(const Color& c1, const Color& c2);
 		friend bool operator!=(const Color& c1, const Color& c2);
 
 		Color& operator+=(const Color& c);
 		Color& operator-=(const Color& c);
-		Color& operator*=(float s);
-		Color& operator/=(float s);
+		Color& operator*=(double s);
+		Color& operator/=(double s);
 
 		uint32_t getAbgrValue() const;
 		bool isTransparent() const;
 		void clamp();
 		Color clamped() const;
 
-		static Color lerp(const Color& start, const Color& end, float alpha);
+		static Color lerp(const Color& start, const Color& end, double alpha);
 		static Color alphaBlend(const Color& first, const Color& second);
 		static uint32_t alphaBlend(uint32_t first, uint32_t second);
 
@@ -45,9 +45,9 @@ namespace Raycer
 		static const Color WHITE;
 		static const Color BLACK;
 
-		float r;
-		float g;
-		float b;
-		float a;
+		double r;
+		double g;
+		double b;
+		double a;
 	};
 }
